refactor: streamline edit slack agent
